+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 * feat: support text input from clipboard (https://github.com/zellij-org/zellij/pull/1926)
+* errors: Don't log errors from panes when quitting zellij (https://github.com/zellij-org/zellij/pull/1918)
 
 ## [0.33.0] - 2022-11-10